Refactor Stripe webhook handling
