+++ app/services/stripe_webhook/base_handler.rb
@@ -1,0 +1,19 @@
+module StripeWebhook
+  class BaseHandler
+    attr_reader :event
+
+    def initialize(event)
+      @event = event
+    end
+
+    def handle
+      raise NotImplementedError
+    end
+
+    private
+
+    def event_object
+      event.data.object
+    end
+  end
+end
