+++ app/services/stripe_webhook/payment_failed_handler.rb
@@ -1,0 +1,11 @@
+module StripeWebhook
+  class PaymentFailedHandler < BaseHandler
+    def handle
+      subscription = Subscription.find_by(stripe_subscription_id: event_object.subscription)
+      return unless subscription
+
+      Rails.logger.warn "Payment failed for subscription #{subscription.id}"
+      # Future: send notification email to user
+    end
+  end
+end
