+++ app/services/stripe_webhook/subscription_deleted_handler.rb
@@ -1,0 +1,10 @@
+module StripeWebhook
+  class SubscriptionDeletedHandler < BaseHandler
+    def handle
+      subscription = Subscription.find_by(stripe_subscription_id: event_object.id)
+      return unless subscription
+
+      subscription.update!(status: "canceled")
+    end
+  end
+end
